Various changes and fixes to Pilot coil components in android
Allow image builder to be customized externally in PilotResizableRemoteImage
Make sure AsyncImagePainter transform does not override image with placeholder when loading from cache.
Fix an issue in PilotResizableRemoteImage where the provided modifier was applied twice (once on the BoxWithContraints and once on the AsyncImage). Instead of doing that we now match parent size on the AsyncImage modifier.